Feat/#59: 에러 핸들링을 위한 응답 메시지 생성

+++ src/main/java/com/poorlex/poorlex/weeklybudget/domain/WeeklyBudget.java
@@ -1,14 +1,11 @@
 package com.poorlex.poorlex.weeklybudget.domain;
 
-import jakarta.persistence.Embedded;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import java.time.LocalDateTime;
-import java.time.temporal.ChronoUnit;
+import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -47,7 +44,7 @@ public class WeeklyBudget {
         return id;
     }
 
-    public int getAmount() {
+    public Long getAmount() {
         return amount.getValue();
     }
