+++ src/main/java/com/poorlex/poorlex/weeklybudget/service/dto/response/WeeklyBudgetLeftResponse.java
@@ -1,9 +1,10 @@
 package com.poorlex.poorlex.weeklybudget.service.dto.response;
 
 import com.poorlex.poorlex.weeklybudget.domain.WeeklyBudget;
-import java.util.Objects;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+
+import java.util.Objects;
 
 @RequiredArgsConstructor
 @Getter
@@ -11,15 +12,15 @@ public class WeeklyBudgetLeftResponse {
 
     private final boolean exist;
 
-    private final int amount;
+    private final Long amount;
 
     public static WeeklyBudgetLeftResponse withNullWeeklyBudget() {
-        return new WeeklyBudgetLeftResponse(false, 0);
+        return new WeeklyBudgetLeftResponse(false, 0L);
     }
 
     public static WeeklyBudgetLeftResponse from(final WeeklyBudget weeklyBudget, final int sumExpenditures) {
         if (Objects.isNull(weeklyBudget)) {
-            return new WeeklyBudgetLeftResponse(false, 0);
+            return new WeeklyBudgetLeftResponse(false, 0L);
         }
         return new WeeklyBudgetLeftResponse(true, weeklyBudget.getAmount() - sumExpenditures);
     }
